rename a python file
