+++ foreshadow/intents/neither.py
